+++ lib/PLS.py
@@ -5,117 +5,64 @@ from scipy.optimize import root
 from scipy.optimize import minimize
 import scipy.sparse
 import os
-import time
-import matplotlib.pyplot as plt
-from matplotlib.pyplot import imshow
-import sparse
-import nibabel as nib
-import nilearn
-import sparse
 from lib.tools2d import *
 
 def PLS2D(theta, ZtX, ZtY, XtX, ZtZ, XtY, YtX, YtZ, XtZ, YtY, n, P, I, tinds, rinds, cinds):
 
     # Obtain Lambda
-    #t1 = time.time()
     Lambda = mapping2D(theta, tinds, rinds, cinds)
-    #t2 = time.time()
-    #print(t2-t1)#3.170967102050781e-05   9
     
     # Obtain Lambda'
-    #t1 = time.time()
     Lambdat = spmatrix.trans(Lambda)
-    #t2 = time.time()
-    #print(t2-t1)# 3.5762786865234375e-06
 
     # Obtain Lambda'Z'Y and Lambda'Z'X
-    #t1 = time.time()
     LambdatZtY = Lambdat*ZtY
     LambdatZtX = Lambdat*ZtX
-    #t2 = time.time()
-    #print(t2-t1)#1.049041748046875e-05   13
     
     # Obtain the cholesky decomposition
-    #t1 = time.time()
     LambdatZtZLambda = Lambdat*(ZtZ*Lambda)
-    #t2 = time.time()
-    #print(t2-t1)#3.790855407714844e-05   2
-    
-    #t1 = time.time()
     chol_dict = sparse_chol2D(LambdatZtZLambda+I, perm=P, retF=True, retP=False, retL=False)
     F = chol_dict['F']
-    #t2 = time.time()
-    #print(t2-t1)#0.0001342296600341797   1
 
     # Obtain C_u (annoyingly solve writes over the second argument,
     # whereas spsolve outputs)
-    #t1 = time.time()
     Cu = LambdatZtY[P,:]
     cholmod.solve(F,Cu,sys=4)
-    #t2 = time.time()
-    #print(t2-t1)#1.5974044799804688e-05   5
 
     # Obtain RZX
-    #t1 = time.time()
     RZX = LambdatZtX[P,:]
     cholmod.solve(F,RZX,sys=4)
-    #t2 = time.time()
-    #print(t2-t1)#1.2159347534179688e-05   7
 
     # Obtain RXtRX
-    #t1 = time.time()
     RXtRX = XtX - matrix.trans(RZX)*RZX
-    #t2 = time.time()
-    #print(t2-t1)#9.775161743164062e-06  11
 
     # Obtain beta estimates (note: gesv also replaces the second
     # argument)
-    #t1 = time.time()
     betahat = XtY - matrix.trans(RZX)*Cu
     try:
         lapack.posv(RXtRX, betahat)
     except:
         lapack.gesv(RXtRX, betahat)
-    #t2 = time.time()
-    #print(t2-t1)#1.7404556274414062e-05   6
 
     # Obtain u estimates
-    #t1 = time.time()
     uhat = Cu-RZX*betahat
     cholmod.solve(F,uhat,sys=5)
     cholmod.solve(F,uhat,sys=8)
-    #t2 = time.time()
-    #print(t2-t1)#1.2874603271484375e-05   8
 
     # Obtain b estimates
-    #t1 = time.time()
     bhat = Lambda*uhat
-    #t2 = time.time()
-    #print(t2-t1)#2.86102294921875e-06  15
     
     # Obtain residuals sum of squares
-    #t1 = time.time()
     resss = YtY-2*YtX*betahat-2*YtZ*bhat+2*matrix.trans(betahat)*XtZ*bhat+matrix.trans(betahat)*XtX*betahat+matrix.trans(bhat)*ZtZ*bhat
-    #t2 = time.time()
-    #print(t2-t1)#3.409385681152344e-05   4
     
     # Obtain penalised residual sum of squares
-    #t1 = time.time()
     pss = resss + matrix.trans(uhat)*uhat
-    #t2 = time.time()
-    #print(t2-t1)#2.6226043701171875e-06  16
     
     # Obtain Log(|L|^2)
-    #t1 = time.time()
     logdet = 2*sum(cvxopt.log(cholmod.diag(F)))
-    #t2 = time.time()
-    #print(t2-t1)#1.5735626220703125e-05   14
     
     # Obtain log likelihood
-    #t1 = time.time()
     logllh = -logdet/2-n/2*(1+np.log(2*np.pi*pss[0,0])-np.log(n))
-    #t2 = time.time()
-    #print(t2-t1)#4.506111145019531e-05   3
     
     return(-logllh)
