+++ lib/blmm_estimate.py
@@ -3,22 +3,11 @@ import warnings as w
 # be ignored for now.
 w.simplefilter(action = 'ignore', category = FutureWarning)
 import numpy as np
-import subprocess
-import warnings
-import resource
-import nibabel as nib
-import sys
 import os
-import glob
-import shutil
-import yaml
-import time
-import warnings
-import subprocess
 np.set_printoptions(threshold=np.nan)
 from scipy import stats
-from lib.tools3d import *
-from lib.tools2d import *
+from lib.npMatrix3d import *
+from lib.npMatrix2d import *
 from lib.fileio import *
 from lib.est3D import *
